refactor: move window content to one dir
